Make per-site configuration more centralized
Create config/sites.yml and load it into SITE_CONFIG (and define
SITE_KIND, which indicates which SITE_CONFIG entry was loaded).

Migrate all the configuration to SITE_CONFIG, which includes making
the title configurable.

This massively cuts down on environment variables - set BOARD_SITE in
the environment to pick a configuration, :devel is used if you don't
specify.