Disable caching on recently replied to threads
This commit bluntly fixes the problem where now post indicators are
stale.  It does so by saying that a thread is not cached if someone has
replied to it in the last 24 hours. This will increase load times
somewhat, though I suspect it's unlikely more than a fifth of threads
are being replied to at any given time, and that at least some of them
are small, and so the problems will be minor in practice.